requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
opencv-python-headless==4.8.1.78
numpy==1.24.3
Pillow==10.1.0
